Move suggestion logging into VS Code specifc part and log after we retrieve from cache
